% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-getTaxAmt.R
\name{getTaxAmt}
\alias{getTaxAmt}
\title{获取税额}
\usage{
getTaxAmt(x = 3000, data = NULL, version = "201810")
}
\arguments{
\item{x}{数据}

\item{data}{税率表数据}

\item{version}{版本号}
}
\value{
返回值
}
\description{
获取税额
}
\examples{
getTaxAmt(4000);
}
