% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-getTaxAmt.R
\name{getTaxInfo}
\alias{getTaxInfo}
\title{计算个人信息}
\usage{
getTaxInfo(x = 3000, data = NULL, version = "201810")
}
\arguments{
\item{x}{数据}

\item{data}{税率表信息}

\item{version}{版本信息，默认为201810版}
}
\value{
返回值
}
\description{
计算个人信息
}
\examples{
getTaxInfo(x)$taxAmount;
}
