% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-getTaxAmt.R
\name{getTaxLevel}
\alias{getTaxLevel}
\title{获取个税等级信息}
\usage{
getTaxLevel(x = 3000, data = NULL, version = "201810")
}
\arguments{
\item{x}{应税所得}

\item{data}{税率表数据}

\item{version}{版本号}
}
\value{
返回值
}
\description{
获取个税等级信息
}
\examples{
getTaxLevel(3000);
}
