% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-getTaxAmt.R
\name{getTaxOutput}
\alias{getTaxOutput}
\title{获取个税结果信息}
\usage{
getTaxOutput(preTaxAmt, threshold = 5000)
}
\arguments{
\item{preTaxAmt}{税前收入}

\item{threshold}{起征点}
}
\value{
返回值
}
\description{
获取个税结果信息
}
\examples{
getTaxOutput(300:400*10);
}
