% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-getTaxAmt.R
\name{getTaxRate}
\alias{getTaxRate}
\title{获取税率}
\usage{
getTaxRate(x = 3000, data = NULL, version = "201810")
}
\arguments{
\item{x}{数据}

\item{data}{税率表}

\item{version}{默认版本201810}
}
\value{
返回值
}
\description{
获取税率
}
\examples{
getTaxRate(3000);
}
