% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-defineClass.R
\docType{class}
\name{taxRateTable-class}
\alias{taxRateTable-class}
\title{计算税率表}
\value{
返回值
}
\description{
计算税率表
}
\section{Slots}{

\describe{
\item{\code{FLevel}}{Integer. 等级}

\item{\code{start}}{numeric.  开始金额}

\item{\code{end}}{numeric.  结束金额}

\item{\code{taxRate}}{integer.  税率}

\item{\code{deductAmt}}{numeric. 扣除数}
}}

\examples{
taxRateTable();
}
