% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-defineClass.R
\name{taxRateTable}
\alias{taxRateTable}
\title{定义相关的函数}
\usage{
taxRateTable(data, version = "201810")
}
\arguments{
\item{data}{原始数据，数据框}

\item{version}{定义版本}
}
\value{
返回值
}
\description{
定义相关的函数
}
\examples{
taxRateTable(data);
}
