% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-tableCalc.R
\name{writePTCIntoTable}
\alias{writePTCIntoTable}
\title{将个税数据写入工资表}
\usage{
writePTCIntoTable(data)
}
\arguments{
\item{data}{符合模板的个税表数据}
}
\value{
返回值
}
\description{
将个税数据写入工资表
}
\examples{
writePTCIntoTable(mydata);
}
